# gateway-service/api/gateway/v1/gateway_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.12.3
// source: gateway/v1/gateway.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGatewayServiceGetGameApp = "/gateway.v1.GatewayService/GetGameApp"
const OperationGatewayServiceGetGatewayInfo = "/gateway.v1.GatewayService/GetGatewayInfo"
const OperationGatewayServiceGetUserGameAppStats = "/gateway.v1.GatewayService/GetUserGameAppStats"
const OperationGatewayServiceListUsersWithPage = "/gateway.v1.GatewayService/ListUsersWithPage"
const OperationGatewayServiceLogin = "/gateway.v1.GatewayService/Login"

type GatewayServiceHTTPServer interface {
	GetGameApp(context.Context, *GetGameAppRequest) (*GetGameAppReply, error)
	// GetGatewayInfo 获取网关信息
	GetGatewayInfo(context.Context, *GetUserRequest) (*GetUserReply, error)
	// GetUserGameAppStats用户和游戏统计接口
	GetUserGameAppStats(context.Context, *GetUserGameAppStatsRequest) (*GetUserGameAppStatsReply, error)
	// ListUsersWithPage分页查询用户列表
	// post body json格式，参数放在请求体中
	ListUsersWithPage(context.Context, *ListUsersRequest) (*ListUsersReply, error)
	// Login 用户邮箱密码登录
	Login(context.Context, *LoginRequest) (*LoginReply, error)
}

func RegisterGatewayServiceHTTPServer(s *http.Server, srv GatewayServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/login", _GatewayService_Login0_HTTP_Handler(srv))
	r.GET("/v1/users/{id}", _GatewayService_GetGatewayInfo0_HTTP_Handler(srv))
	r.POST("/v1/users", _GatewayService_ListUsersWithPage0_HTTP_Handler(srv))
	r.GET("/v1/game_app/{id}", _GatewayService_GetGameApp0_HTTP_Handler(srv))
	r.GET("/v1/user_game_app_stats", _GatewayService_GetUserGameAppStats0_HTTP_Handler(srv))
}

func _GatewayService_Login0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _GatewayService_GetGatewayInfo0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceGetGatewayInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGatewayInfo(ctx, req.(*GetUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserReply)
		return ctx.Result(200, reply)
	}
}

func _GatewayService_ListUsersWithPage0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListUsersRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceListUsersWithPage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListUsersWithPage(ctx, req.(*ListUsersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListUsersReply)
		return ctx.Result(200, reply)
	}
}

func _GatewayService_GetGameApp0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetGameAppRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceGetGameApp)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGameApp(ctx, req.(*GetGameAppRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetGameAppReply)
		return ctx.Result(200, reply)
	}
}

func _GatewayService_GetUserGameAppStats0_HTTP_Handler(srv GatewayServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserGameAppStatsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayServiceGetUserGameAppStats)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserGameAppStats(ctx, req.(*GetUserGameAppStatsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserGameAppStatsReply)
		return ctx.Result(200, reply)
	}
}

type GatewayServiceHTTPClient interface {
	GetGameApp(ctx context.Context, req *GetGameAppRequest, opts ...http.CallOption) (rsp *GetGameAppReply, err error)
	// GetGatewayInfo 获取网关信息
	GetGatewayInfo(ctx context.Context, req *GetUserRequest, opts ...http.CallOption) (rsp *GetUserReply, err error)
	// GetUserGameAppStats用户和游戏统计接口
	GetUserGameAppStats(ctx context.Context, req *GetUserGameAppStatsRequest, opts ...http.CallOption) (rsp *GetUserGameAppStatsReply, err error)
	// ListUsersWithPage分页查询用户列表
	// post body json格式，参数放在请求体中
	ListUsersWithPage(ctx context.Context, req *ListUsersRequest, opts ...http.CallOption) (rsp *ListUsersReply, err error)
	// Login 用户邮箱密码登录
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
}

type GatewayServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewGatewayServiceHTTPClient(client *http.Client) GatewayServiceHTTPClient {
	return &GatewayServiceHTTPClientImpl{client}
}

func (c *GatewayServiceHTTPClientImpl) GetGameApp(ctx context.Context, in *GetGameAppRequest, opts ...http.CallOption) (*GetGameAppReply, error) {
	var out GetGameAppReply
	pattern := "/v1/game_app/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGatewayServiceGetGameApp))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGatewayInfo 获取网关信息
func (c *GatewayServiceHTTPClientImpl) GetGatewayInfo(ctx context.Context, in *GetUserRequest, opts ...http.CallOption) (*GetUserReply, error) {
	var out GetUserReply
	pattern := "/v1/users/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGatewayServiceGetGatewayInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUserGameAppStats用户和游戏统计接口
func (c *GatewayServiceHTTPClientImpl) GetUserGameAppStats(ctx context.Context, in *GetUserGameAppStatsRequest, opts ...http.CallOption) (*GetUserGameAppStatsReply, error) {
	var out GetUserGameAppStatsReply
	pattern := "/v1/user_game_app_stats"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGatewayServiceGetUserGameAppStats))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListUsersWithPage分页查询用户列表
// post body json格式，参数放在请求体中
func (c *GatewayServiceHTTPClientImpl) ListUsersWithPage(ctx context.Context, in *ListUsersRequest, opts ...http.CallOption) (*ListUsersReply, error) {
	var out ListUsersReply
	pattern := "/v1/users"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGatewayServiceListUsersWithPage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Login 用户邮箱密码登录
func (c *GatewayServiceHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/v1/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGatewayServiceLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}